+++ pteroxide-models/src/application/allocation.rs
@@ -5,7 +5,7 @@ pub struct Allocation {
     pub id: i32,
     pub ip: String,
     pub alias: Option<String>,
-    pub port: i64,
+    pub port: i32,
     pub notes: Option<String>,
     pub assigned: bool,
 }
